Module Private_.Plugins

module type S = sig ... end

The signature of the modules present in the module generated by the stanza generate_sites_module

module Make (_ : sig ... end) : S

The functor applied in the generated module

val load : string -> unit

Load a library

val available : string -> bool

Indicates if a library exists in the search path